refactor: 종속되는 dto 리팩토링
