+++ backend/src/main/java/site/kkokkio/domain/member/controller/MemberController.java
@@ -10,10 +10,10 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import site.kkokkio.domain.member.dto.MemberLoginRequest;
-import site.kkokkio.domain.member.dto.MemberLoginResponse;
-import site.kkokkio.domain.member.dto.MemberResponse;
-import site.kkokkio.domain.member.dto.MemberSignUpRequest;
+import site.kkokkio.domain.member.controller.dto.MemberLoginRequest;
+import site.kkokkio.domain.member.controller.dto.MemberLoginResponse;
+import site.kkokkio.domain.member.controller.dto.MemberResponse;
+import site.kkokkio.domain.member.controller.dto.MemberSignUpRequest;
 import site.kkokkio.domain.member.service.MemberService;
 import site.kkokkio.global.dto.RsData;
 import site.kkokkio.global.util.JwtUtils;
