+++ backend/src/main/java/site/kkokkio/domain/member/controller/dto/MemberLoginResponse.java
@@ -1,7 +1,6 @@
-package site.kkokkio.domain.member.dto;
+package site.kkokkio.domain.member.controller.dto;
 
 import java.time.LocalDateTime;
-import java.util.UUID;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
@@ -9,7 +8,6 @@ import site.kkokkio.domain.member.entity.Member;
 import site.kkokkio.global.enums.MemberRole;
 
 public record MemberLoginResponse(
-	UUID id,
 	String nickname,
 	String email,
 	LocalDateTime deleteAt,
@@ -18,7 +16,7 @@ public record MemberLoginResponse(
 	String token
 ) {
 	public static MemberLoginResponse of(Member member, String token) {
-		return new MemberLoginResponse(member.getId(), member.getNickname(), member.getEmail(),
+		return new MemberLoginResponse(member.getNickname(), member.getEmail(),
 			member.getDeletedAt(), member.getRole(), token);
 	}
 }
