+++ backend/src/main/java/site/kkokkio/domain/member/controller/dto/MemberResponse.java
@@ -1,4 +1,4 @@
-package site.kkokkio.domain.member.dto;
+package site.kkokkio.domain.member.controller.dto;
 
 import java.time.LocalDate;
